Fix paper compatibility (#645)
* Update libraries

* Add temporary fix for paper

* Update to java 17

* Rename 1.18 package

* Rename 1.19 and 1.20 packages

* Switch from spigot to paper

* Delete Entities.java

* Move to new packages

* Fix worldutils
