+++ modules/Movecraft/pom.xml
@@ -22,29 +22,22 @@
         </pluginRepository>
     </pluginRepositories>
 
-    <repositories>
-        <repository>
-            <id>spigot-repo</id>
-            <url>https://hub.spigotmc.org/nexus/content/repositories/snapshots/</url>
-        </repository>
-    </repositories>
-
     <dependencies>
         <dependency>
             <groupId>net.countercraft</groupId>
-            <artifactId>movecraft-v1_18_r2</artifactId>
+            <artifactId>movecraft-v1_18</artifactId>
             <version>${revision}</version>
             <type>jar</type>
         </dependency>
         <dependency>
             <groupId>net.countercraft</groupId>
-            <artifactId>movecraft-v1_19_r3</artifactId>
+            <artifactId>movecraft-v1_19</artifactId>
             <version>${revision}</version>
             <type>jar</type>
         </dependency>
         <dependency>
             <groupId>net.countercraft</groupId>
-            <artifactId>movecraft-v1_20_R4</artifactId>
+            <artifactId>movecraft-v1_20</artifactId>
             <version>${revision}</version>
             <type>jar</type>
         </dependency>
@@ -125,8 +118,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.8.1</version>
                 <configuration>
-                    <source>16</source>
-                    <target>16</target>
+                    <source>17</source>
+                    <target>17</target>
                 </configuration>
             </plugin>
             <plugin>
@@ -166,19 +159,19 @@
                             <createDependencyReducedPom>false</createDependencyReducedPom>
                             <filters>
                                 <filter>
-                                    <artifact>net.countercraft:movecraft-v1_18_r2</artifact>
+                                    <artifact>net.countercraft:movecraft-v1_18</artifact>
                                     <includes>
                                         <include>**</include>
                                     </includes>
                                 </filter>
                                 <filter>
-                                    <artifact>net.countercraft:movecraft-v1_19_r3</artifact>
+                                    <artifact>net.countercraft:movecraft-v1_19</artifact>
                                     <includes>
                                         <include>**</include>
                                     </includes>
                                 </filter>
                                 <filter>
-                                    <artifact>net.countercraft:movecraft-v1_20_R4</artifact>
+                                    <artifact>net.countercraft:movecraft-v1_20</artifact>
                                     <includes>
                                         <include>**</include>
                                     </includes>
