+++ modules/Movecraft/src/main/resources/config.yml
@@ -26,5 +26,4 @@ ForbiddenRemoteSigns: # These strings will be ignored when searching for signs t
   - "Mounted"
   - "Denied"
   - "[private]"
-CompatibilityMode: false
 GeneratedDatapack: false # Set to false to re-generate datapack on startup
