+++ modules/api/src/main/java/net/countercraft/movecraft/WorldHandler.java
@@ -15,7 +15,10 @@ public abstract class WorldHandler {
     public abstract void translateCraft(@NotNull Craft craft, @NotNull MovecraftLocation newLocation, @NotNull World world);
     public abstract void setBlockFast(@NotNull Location location, @NotNull BlockData data);
     public abstract void setBlockFast(@NotNull Location location, @NotNull MovecraftRotation rotation, @NotNull BlockData data);
-    public abstract void disableShadow(@NotNull Material type);
     public abstract @Nullable Location getAccessLocation(@NotNull InventoryView inventoryView);
     public abstract void setAccessLocation(@NotNull InventoryView inventoryView, @NotNull Location location);
+
+    public static @NotNull String getPackageName(@NotNull String minecraftVersion) {
+        return "v1_" + minecraftVersion.substring(minecraftVersion.indexOf('.') + 1, minecraftVersion.lastIndexOf('.'));
+    }
 }
