+++ modules/api/src/main/java/net/countercraft/movecraft/support/AsyncChunk.java
@@ -1,6 +1,7 @@
 package net.countercraft.movecraft.support;
 
 import net.countercraft.movecraft.MovecraftLocation;
+import net.countercraft.movecraft.WorldHandler;
 import org.bukkit.Bukkit;
 import org.bukkit.Chunk;
 import org.bukkit.Material;
@@ -15,17 +16,15 @@ public abstract class AsyncChunk<T extends Chunk> {
 
     private static final Constructor<?> constructor;
     static {
-        String packageName = Bukkit.getServer().getClass().getPackage().getName();
-        String version = packageName.substring(packageName.lastIndexOf('.') + 1);
         Constructor<?> temp = null;
         try {
-            Class.forName("net.countercraft.movecraft.support.v1_18_R2.IAsyncChunk");
-            final Class<?> clazz = Class.forName("net.countercraft.movecraft.support." + version + ".IAsyncChunk");
+            Class.forName("net.countercraft.movecraft.support.v1_18.IAsyncChunk");
+            final Class<?> clazz = Class.forName("net.countercraft.movecraft.support." + WorldHandler.getPackageName(Bukkit.getServer().getMinecraftVersion()) + ".IAsyncChunk");
             if (AsyncChunk.class.isAssignableFrom(clazz)) {
                 temp = clazz.getConstructor(Chunk.class);
             }
         } catch (ClassNotFoundException | NoSuchMethodException | SecurityException exception) {
-            Bukkit.getLogger().severe(String.format("Error in registering Chunk accessor for version %s from the classpath.", version));
+            Bukkit.getLogger().severe(String.format("Error in registering Chunk accessor for version %s from the classpath.", Bukkit.getServer().getMinecraftVersion()));
             exception.printStackTrace();
         }
         constructor = temp;
