+++ pom.xml
@@ -14,9 +14,9 @@
 
     <modules>
         <module>modules/api</module>
-        <module>modules/v1_18_R2</module>
-        <module>modules/v1_19_R3</module>
-        <module>modules/v1_20_R4</module>
+        <module>modules/v1_18</module>
+        <module>modules/v1_19</module>
+        <module>modules/v1_20</module>
         <module>modules/datapack</module>
         <module>modules/Movecraft</module>
     </modules>
@@ -35,8 +35,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.8.1</version>
                 <configuration>
-                    <source>13</source>
-                    <target>13</target>
+                    <source>17</source>
+                    <target>17</target>
                 </configuration>
             </plugin>
             <plugin>
